+++ lib/langchain/llm/ollama.rb
@@ -5,21 +5,24 @@ module Langchain::LLM
   # Available models: https://ollama.ai/library
   #
   # Usage:
-  #    ollama = Langchain::LLM::Ollama.new(url: ENV["OLLAMA_URL"])
+  #    ollama = Langchain::LLM::Ollama.new(url: ENV["OLLAMA_URL"], default_options: {})
   #
   class Ollama < Base
-    attr_reader :url
+    attr_reader :url, :defaults
 
     DEFAULTS = {
-      temperature: 0.0,
+      temperature: 0.8,
       completion_model_name: "llama2",
       embeddings_model_name: "llama2"
     }.freeze
 
     # Initialize the Ollama client
     # @param url [String] The URL of the Ollama instance
-    def initialize(url:)
+    # @param default_options [Hash] The default options to use
+    #
+    def initialize(url:, default_options: {})
       @url = url
+      @defaults = DEFAULTS.merge(default_options)
     end
 
     #
@@ -27,32 +30,87 @@ module Langchain::LLM
     #
     # @param prompt [String] The prompt to complete
     # @param model [String] The model to use
-    # @param options [Hash] The options to use (https://github.com/jmorganca/ollama/blob/main/docs/modelfile.md#valid-parameters-and-values)
+    #   For a list of valid parameters and values, see:
+    #   https://github.com/jmorganca/ollama/blob/main/docs/modelfile.md#valid-parameters-and-values
     # @return [Langchain::LLM::OllamaResponse] Response object
     #
-    def complete(prompt:, model: nil, **options)
-      response = +""
+    def complete(
+      prompt:,
+      model: defaults[:completion_model_name],
+      images: nil,
+      format: nil,
+      system: nil,
+      template: nil,
+      context: nil,
+      stream: nil,
+      raw: nil,
+      mirostat: nil,
+      mirostat_eta: nil,
+      mirostat_tau: nil,
+      num_ctx: nil,
+      num_gqa: nil,
+      num_gpu: nil,
+      num_thread: nil,
+      repeat_last_n: nil,
+      repeat_penalty: nil,
+      temperature: defaults[:temperature],
+      seed: nil,
+      stop: nil,
+      tfs_z: nil,
+      num_predict: nil,
+      top_k: nil,
+      top_p: nil,
+      &block
+    )
 
-      model_name = model || DEFAULTS[:completion_model_name]
+      parameters = {
+        prompt: prompt,
+        model: model,
+        images: images,
+        format: format,
+        system: system,
+        template: template,
+        context: context,
+        stream: stream,
+        raw: raw
+      }.compact
+
+      llm_parameters = {
+        mirostat: mirostat,
+        mirostat_eta: mirostat_eta,
+        mirostat_tau: mirostat_tau,
+        num_ctx: num_ctx,
+        num_gqa: num_gqa,
+        num_gpu: num_gpu,
+        num_thread: num_thread,
+        repeat_last_n: repeat_last_n,
+        repeat_penalty: repeat_penalty,
+        temperature: temperature,
+        seed: seed,
+        stop: stop,
+        tfs_z: tfs_z,
+        num_predict: num_predict,
+        top_k: top_k,
+        top_p: top_p
+      }
+
+      parameters[:options] = llm_parameters.compact
+
+      response = ""
 
       client.post("api/generate") do |req|
-        req.body = {}
-        req.body["prompt"] = prompt
-        req.body["model"] = model_name
+        req.body = parameters
 
-        req.body["options"] = options if options.any?
-
-        # TODO: Implement streaming support when a &block is passed in
         req.options.on_data = proc do |chunk, size|
           json_chunk = JSON.parse(chunk)
 
-          unless json_chunk.dig("done")
-            response.to_s << JSON.parse(chunk).dig("response")
-          end
+          response += json_chunk.dig("response")
+
+          yield json_chunk, size if block
         end
       end
 
-      Langchain::LLM::OllamaResponse.new(response, model: model_name)
+      Langchain::LLM::OllamaResponse.new(response, model: parameters[:model])
     end
 
     #
@@ -63,18 +121,57 @@ module Langchain::LLM
     # @param options [Hash] The options to use
     # @return [Langchain::LLM::OllamaResponse] Response object
     #
-    def embed(text:, model: nil, **options)
-      model_name = model || DEFAULTS[:embeddings_model_name]
+    def embed(
+      text:,
+      model: defaults[:embeddings_model_name],
+      mirostat: nil,
+      mirostat_eta: nil,
+      mirostat_tau: nil,
+      num_ctx: nil,
+      num_gqa: nil,
+      num_gpu: nil,
+      num_thread: nil,
+      repeat_last_n: nil,
+      repeat_penalty: nil,
+      temperature: nil,
+      seed: nil,
+      stop: nil,
+      tfs_z: nil,
+      num_predict: nil,
+      top_k: nil,
+      top_p: nil
+    )
+      parameters = {
+        prompt: text,
+        model: model
+      }.compact
+
+      llm_parameters = {
+        mirostat: mirostat,
+        mirostat_eta: mirostat_eta,
+        mirostat_tau: mirostat_tau,
+        num_ctx: num_ctx,
+        num_gqa: num_gqa,
+        num_gpu: num_gpu,
+        num_thread: num_thread,
+        repeat_last_n: repeat_last_n,
+        repeat_penalty: repeat_penalty,
+        temperature: temperature,
+        seed: seed,
+        stop: stop,
+        tfs_z: tfs_z,
+        num_predict: num_predict,
+        top_k: top_k,
+        top_p: top_p
+      }
+
+      parameters[:options] = llm_parameters.compact
 
       response = client.post("api/embeddings") do |req|
-        req.body = {}
-        req.body["prompt"] = text
-        req.body["model"] = model_name
-
-        req.body["options"] = options if options.any?
+        req.body = parameters
       end
 
-      Langchain::LLM::OllamaResponse.new(response.body, model: model_name)
+      Langchain::LLM::OllamaResponse.new(response.body, model: parameters[:model])
     end
 
     private