+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+- Fix and refactor of `Langchain::LLM::Ollama`, responses can now be streamed.
 
 ## [0.9.0]
 - Introducing new `Langchain::Assistant` that will be replacing `Langchain::Conversation` and `Langchain::Agent`s.